cleanup main.js
