+++ gui/src/main/fileutils.js
@@ -1,0 +1,29 @@
+
+module.exports.selectIcon = (info) => {
+  if (['image', 'video'].indexOf(info.class) !== -1) {
+    return info.class
+  } else if (info.class === 'music') {
+    return 'audio'
+  } else if (info.mime === 'application/pdf') {
+    return 'pdf'
+  } else if (info.mime === 'application/x-binary') {
+    return 'binary'
+  } else if (!info.mime) {
+    return 'file'
+  } else if (info.mime.match(/[/-][bg]?zip2?$/)) {
+    return 'archive'
+  } else if (info.mime.match(/^(text|application)\/(html|xml)/)) {
+    return 'code'
+  } else if (info.mime.match(/^text\//)) {
+    return 'text'
+  } else if (info.mime.match(/^application\/.*rtf/)) {
+    return 'text'
+  } else if (info.mime.match(/word/)) {
+    return 'text'
+  } else if (info.mime.match(/powerpoint/)) {
+    return 'presentation'
+  } else if (info.mime.match(/excel/)) {
+    return 'spreadsheet'
+  }
+  return 'file'
+}
